raspberry-pi-gui/requirements.txt
pyserial
numpy
scipy
matplotlib
pynanovna
pandas